timer on spring boot update

+++ src/main/java/org/jibe77/hermanas/controller/music/MusicController.java
@@ -7,6 +7,7 @@ import org.springframework.beans.factory.config.ConfigurableBeanFactory;
 import org.springframework.context.annotation.Scope;
 import org.springframework.stereotype.Component;
 
+import javax.annotation.PreDestroy;
 import java.io.*;
 import java.security.SecureRandom;
 import java.util.*;
@@ -46,9 +47,17 @@ public class MusicController {
     @Value("${music.volume.regular}")
     private String volumeLevelRegular;
 
+    @Value("${music.security.timer.delay}")
+    private long musicSecurityTimerDelay;
+
+    @Value("${music.enabled}")
+    private boolean musicEnabled;
+
     ProcessLauncher processLauncher;
 
     private Process currentMusicProcess;
+
+    private Timer musicSecurityStopTimer;
 
     Logger logger = LoggerFactory.getLogger(MusicController.class);
 
@@ -57,16 +66,20 @@ public class MusicController {
     }
 
     public boolean playMusicRandomly() {
-        stop();
-        try {
-            setMusicLevel(volumeLevelRegular);
-            List<String> listOfFile = getListOfFiles(pathToFolder);
-            playMusic(listOfFile);
-        } catch (IOException e) {
-            logger.error("Can't play music.", e);
+        if (musicEnabled) {
+            stop();
+            try {
+                setMusicLevel(volumeLevelRegular);
+                List<String> listOfFile = getListOfFiles(pathToFolder);
+                playMusic(listOfFile);
+            } catch (IOException e) {
+                logger.error("Can't play music.", e);
+                return false;
+            }
+            return true;
+        } else {
             return false;
         }
-        return true;
     }
 
     private void playMusic(File musicFile) throws IOException {
@@ -79,13 +92,14 @@ public class MusicController {
         logger.info("Play music with command {} {} {}  {}.",
                 musicPlayerStartCmd, musicPlayerNoDispParam
                 , musicPlayerShuffle, listOfFile);
-        List<String> commandWithParams = new ArrayList<>(listOfFile.size() + 2);
+        List<String> commandWithParams = new ArrayList<>(listOfFile.size() + 3);
         commandWithParams.add(musicPlayerStartCmd);
         commandWithParams.add(musicPlayerNoDispParam);
         commandWithParams.add(musicPlayerShuffle);
         commandWithParams.addAll(listOfFile);
         currentMusicProcess = processLauncher.launch(commandWithParams);
         printErrorStreamInThread(currentMusicProcess);
+        startSecurityTimer();
     }
 
     private void printErrorStreamInThread(Process currentMusicProcess) {
@@ -120,40 +134,73 @@ public class MusicController {
                 .map(File::getAbsolutePath).collect(Collectors.toList());
     }
 
+    private void startSecurityTimer() {
+        if (musicSecurityStopTimer != null) {
+            musicSecurityStopTimer.cancel();
+        }
+        musicSecurityStopTimer = new Timer("Music security stop");
+        musicSecurityStopTimer.schedule(new TimerTask() {
+                                            public void run() {
+                                                logger.info("stopping music after {} ms.", musicSecurityTimerDelay);
+                                                stop();
+                                            }
+                                        },
+                musicSecurityTimerDelay);
+    }
+
     private File pickSong(File[] array) {
         int rnd = new SecureRandom().nextInt(array.length);
         return array[rnd];
     }
 
     public void stop() {
-        if (currentMusicProcess != null) {
+        if (musicEnabled && currentMusicProcess != null) {
             logger.info("Stop music destroying process.");
             currentMusicProcess.destroyForcibly();
             currentMusicProcess = null;
+            if (musicSecurityStopTimer != null) {
+                musicSecurityStopTimer.cancel();
+                musicSecurityStopTimer = null;
+            }
         }
     }
 
     public boolean cocorico() {
-        logger.info("Play cocorico !");
-        stop();
-        try {
-            setMusicLevel(volumeLevelMax);
-            File mixFolder = new File(pathToRooster);
-            File[] filesAvailable = mixFolder.listFiles();
-            File pickedFile = pickSong(filesAvailable);
-            playMusic(pickedFile);
-            return true;
-        } catch (IOException e) {
-            logger.error("Can't play cocorico.", e);
+        if (musicEnabled) {
+            logger.info("Play cocorico !");
+            stop();
+            try {
+                setMusicLevel(volumeLevelMax);
+                File mixFolder = new File(pathToRooster);
+                File[] filesAvailable = mixFolder.listFiles();
+                File pickedFile = pickSong(filesAvailable);
+                playMusic(pickedFile);
+                return true;
+            } catch (IOException e) {
+                logger.error("Can't play cocorico.", e);
+                return false;
+            }
+        } else {
             return false;
+        }
+    }
+
+    @PreDestroy
+    private void tearDown() {
+        if (musicEnabled) {
+            stop();
         }
     }
 
     /**
      * This methods returns true if music is playing
+     *
      * @return true if music is playing
      */
     public boolean isPlaying() {
+        logger.info("status of player is request, current process is null : {} and is alive : {}",
+                currentMusicProcess != null,
+                currentMusicProcess != null && currentMusicProcess.isAlive());
         return (currentMusicProcess != null && currentMusicProcess.isAlive());
     }
